Added stub test for sending email with nodemailer stub transport
