+++ package.json
@@ -7,7 +7,7 @@
     "test": "env NODE_ENV=test mocha --recursive tests/tests",
     "cover": "env NODE_ENV=test istanbul cover _mocha -- --recursive tests/tests",
     "coveralls": "npm run cover -- --report lcovonly && cat ./coverage/lcov.info | coveralls",
-    "start": "node ./app"
+    "start": "env NODE_ENV=dev node ./app"
   },
   "dependencies": {
     "@onehilltech/blueprint": "1.1.1",
@@ -19,6 +19,7 @@
     "mocha": "^3.1.2",
     "mocha-lcov-reporter": "^1.2.0",
     "nodemailer": "^2.6.4",
+    "nodemailer-stub-transport": "^1.1.0",
     "passport": "^0.3.2",
     "passport-http-bearer": "^1.0.1",
     "superagent": "^2.3.0",
